Gruntfile: Optimize verify:source-maps
The `verify:source-maps` check runs serially both to find the set of files to
check as well as to scan those files. Additionally, it loads the entire
contents of those files into memory when running and then performs a full text
search across the contents in memory.

In this patch we're refactoring that check to run asynchronously. It
concurrently calls the `glob` function to search for the files to scan, and
then concurrently scans those files files one 64 KB chunk at a time to look for
the source map text.

This concurrency should eliminate some IO bottlenecking and the chunking might
drop the memory use if the previously loaded files are large enough. Now,
memory use should correspond to roughly 64 KB times the number of concurrently
opened files that are being scanned.